begin memory optimization WIP

+++ opeGL/include/cube.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <glad/glad.h>
 #include <glm/glm.hpp>
 #include <glm/fwd.hpp>
@@ -20,6 +21,6 @@ struct Cube {
   BlockType blockType;
 };
 void setupCubeBuffers(Cube &cube);
-std::vector<GLfloat> cubeVerts(Cube &cube, int xoffset, int yoffset, int zoffset);
+std::vector<GLint> cubeVerts(Cube &cube, int xoffset, int yoffset, int zoffset);
 Cube createCube(size_t diff, size_t spec, glm::vec3 pos, glm::vec3 rotation, float angle, float scale, BlockType blockType = Grass, bool deferBuffers = false);
 void drawCube(Cube &cube, GLuint shader);
